Book 訂正命名

+++ src/main/java/com/tgfc/library/service/imp/BookService.java
@@ -10,16 +10,15 @@ import com.tgfc.library.request.BookDataPageRequest;
 import com.tgfc.library.response.BaseResponse;
 import com.tgfc.library.response.BooksResponse;
 import com.tgfc.library.service.IBookService;
+import com.tgfc.library.service.IPhotoService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -27,17 +26,25 @@ import java.util.List;
 @Service
 public class BookService implements IBookService {
 
-    @Autowired
+
+    private final
     IBookRepository bookDataRepository;
 
-    @Autowired
-    PhotoService photoService;
-    @Autowired
+    private final
+    IPhotoService photoService;
+    private final
     IRecommendRepository iRecommendRepository;
 
 
+    public BookService(IBookRepository bookDataRepository, IPhotoService photoService, IRecommendRepository iRecommendRepository) {
+        this.bookDataRepository = bookDataRepository;
+        this.photoService = photoService;
+        this.iRecommendRepository = iRecommendRepository;
+    }
+
+
     @Override
-    public BaseResponse getBookList(BookDataPageRequest model) throws IOException{
+    public BaseResponse getBookList(BookDataPageRequest model) {
         BaseResponse response = new BaseResponse();
         Pageable pageable = PageRequest.of(model.getPageNumber(), model.getPageSize());
         Page<Book> pageBook = bookDataRepository.findAllByKeyword(model.getKeyword(), pageable);
@@ -61,7 +68,7 @@ public class BookService implements IBookService {
     }
 
     @Override
-    public BaseResponse getById(int id) throws IOException{
+    public BaseResponse getById(int id) {
         BaseResponse response = new BaseResponse();
         Book book = bookDataRepository.getById(id);
         if (book.getPhotoOriginalName() != null && book.getPhotoOriginalName().length() != 0)
@@ -83,7 +90,9 @@ public class BookService implements IBookService {
                 book=new Book();
                 try {
                     BeanUtils.copyProperties(addBook, book);
-                } catch (BeansException e) {}
+                } catch (BeansException e) {
+                    response.setMessage("addBook錯誤");
+                }
                 book=bookDataRepository.save(book);
                 Recommend recommend=iRecommendRepository.findRecommendByIsbn(book.getIsbn());
                 if(recommend!=null)
@@ -104,6 +113,7 @@ public class BookService implements IBookService {
                 addBook.setId(book.getId());
                 BeanUtils.copyProperties(addBook, book);
             } catch (BeansException e) {
+                response.setMessage("Id");
             }
             if (bookDataRepository.save(book) != null)
                 response.setStatus(true);
@@ -117,6 +127,11 @@ public class BookService implements IBookService {
     public BaseResponse deleteBook(Integer id) {
         BaseResponse baseResponse = new BaseResponse();
         Book book=bookDataRepository.getById(id);
+        if(book==null){
+            baseResponse.setMessage("無此ID");
+            return baseResponse;
+        }
+
         bookDataRepository.deleteById(id);
         if(bookDataRepository.getById(id)==null){
             baseResponse.setStatus(true);
